Implement quasiquotes in term of simple eval/exec
Except pyf. It's more complicated
